+++ client/lib/js/easings.js
@@ -111,7 +111,7 @@ var ABeamer;
     });
     function _expressionEasing(t, params, args) {
         ABeamer._vars.t = t;
-        return parseFloat(ABeamer._computeExpression(params.__expression, args));
+        return parseFloat(ABeamer.calcExpr(params.__expression, args));
     }
     ABeamer._expressionEasing = _expressionEasing;
 })(ABeamer || (ABeamer = {}));