+++ client/lib/js/i18n.js
@@ -21,6 +21,7 @@ var ABeamer;
         Msgs["MustNatPositive"] = "The value of %p% must be a natural positive";
         Msgs["MustNatNotNegative"] = "The value of %p% must be a natural non-negative";
         Msgs["MustBeANumber"] = "The value of %p% must be a number";
+        Msgs["MustBeANumberOrExpr"] = "The value of %p% must be a number or an expression";
         Msgs["Unknown"] = "Unknown %p%";
         Msgs["UnknownOf"] = "Unknown \"%type%\" %p%";
         Msgs["UnknownType"] = "Unknown type of %p%";